Turn drucker prager compositions into a lib plugin.
